styles - Added landing page
