+++ src/index.css
@@ -141,21 +141,9 @@
     box-shadow: inset 0 0 0 10em rgb(14, 149, 109);
 }
 
-.title {
-    font-size: 3rem;
-    font-weight: 600;
-    font-family: "Marcellus", Arial, sans-serif;
-    font-weight: bold;
-    text-transform: uppercase;
-    background-image: linear-gradient(to left, #4816d2, #5382e1);
-    color: transparent;
-    background-clip: text;
+.grad {
+    background: linear-gradient(45deg, #b25dea, #5582db);
     -webkit-background-clip: text;
-    text-shadow: 2px 2px 4px rgba(0, 0, 0, 0.4);
-    transition: transform 0.3s ease-in-out;
-    --text-stroke-color: rgba(255, 255, 255, 0.6);
-    --animation-color: #37ff8b;
-    --fs-size: 2em;
-    letter-spacing: 3px;
-    font-size: var(--fs-size);
+    background-clip: text;
+    color: transparent;
 }
